test: Deal with the silent menu in grub
The distro test for EFI in the EFI ARM app is currently flaky.

If the test does not send an 'escape' character, then the board may boot
straight into Ubuntu. If it does, but didn't need to, then grub sits at
the command prompt.

Handle this by pressing escape twice (which should always go to the grub
command line), then using the 'normal' command to start the menu.

Series-to: concept
Series-cc: heinrich

Signed-off-by: Simon Glass <sjg@chromium.org>