expo: Support drawing only the dirty portion of an expo
To optimise drawing on the display, provide a new function which draws
only the part of the display which is marked as dirty.

Series-to: concept
Cover-letter:
expo: Continue development of expo with mouse
This series integrates mouse support into expo and starts to build the
infrastructure for drawing the mouse pointer more efficiently (i.e.
without re-rendering the entire scene).

It also adds a few more structs to video_defs.h for position and size.

This is part D.
END

Signed-off-by: Simon Glass <sjg@chromium.org>