expo: Introduce a test mode
Add a test mode for expo which will display useful debugging
information.

Put this feature behind a CONFIG_EXPO_TEST option to avoid code-size
growth. Create a separate C file and a header. Use static inlines to
avoid lots of CONFIG checking in expo.c

Enable this feature for sandbox and the EFI app for now.

Co-developed-by: Claude <noreply@anthropic.com>
Signed-off-by: Simon Glass <sjg@chromium.org>