expo: Use manual-sync mode
Make use of the manual-sync mode provided by the video subsystem. This
removes uncertainty about whether the display will actually be updated.

Series-to: concept
Series-cc: heinrich
Cover-letter:
expo: More mouse development for expo
This series adds support for hiding the system pointer, if one is
available, since it is confusing to have an expo pointer separate from
the system one.

It also introduces a simple test mode for expo, to check the frame rate,
etc.

Some minor tweaks in the video uclass are also included, as well as
improvements to the video documentation.
END

Signed-off-by: Simon Glass <sjg@chromium.org>