efi: serial: Speed up reading from input
The serial driver is currently quite slow since it sets an event and
waits for at least 1ms for a keypress, which may never come.

On real hardware, the keyboard does not work correctly wihout waiting
for events. Reduce the time to 10us which should be enough to provide an
opportunity for event processing.

QEMU does not actually have this problem, so detect that and skip the
processing altogether.

Co-developed-by: Claude <noreply@anthropic.com>
Signed-off-by: Simon Glass <sjg@chromium.org>