efi: mouse: Move simple-pointer code into a function
In preparation for supporting an absolute pointer, move the setup of
the simple pointer into its own function. Rename 'pointer' to simple
and last_state to simple_last. Take this opportunity to add comments
for struct efi_mouse_priv since it will be less obvious once the new
support is added.

Clean up some unnecessary cruft in the remove() path.

Signed-off-by: Simon Glass <sjg@chromium.org>