expo: Add an image that indicates verification
Where an OS is verified by a vendor, add a way to show a suitable symbol
to indicate this.

Series-to: concept
Series-cc: heinrich
Cover-letter:
expo: Extend the boot menu
The current boot menu works but it is still fairly basic. This series
makes various improvements:

- adds buttons for settings, help
- adds a way to show other items for each OS (e.g. distro version-name)
- adds a second logo
- allows a box to be drawn around each item

This series also provides a way to dump the expo as a text file, so that
it can be examined or compared with another one. This is useful for
debugging. A 'cedit dump' command allows dumping the information.

A few fixes are included as well:

- for mouse operation, avoid dropping click events
- for keyboard, fix support for escape sequences
END

Signed-off-by: Simon Glass <sjg@chromium.org>