boot: Split out setting menu properties from menu creation
The existing menu is functional but basic. It would be useful to be able
to create the objects but apply a different layout or style for them.
The bootflow_menu_new() function is quite long in any case.

Split out the parts which adjust the position, alignment, etc. of the
objects into a separate function. We can (later) call this from
elsewhere to apply the basic styling.

Signed-off-by: Simon Glass <sjg@chromium.org>