expo: Avoid setting SCENEOF_SIZE_VALID with calculated size
When the size of an object is calculated from its contents, we should
not set the SCENEOF_SIZE_VALID flag. This flag prevents the object from
resizing if the font is changed, for example. The flag is intended to
mean that the size was explicitly set by the controller, so we should
generally not set it in the expo implementation.

It is also inefficient to search for the object ID when we already
have the object.

Update scene_set_default_bbox() to just set the size and request that
it be synced.

Signed-off-by: Simon Glass <sjg@chromium.org>