boot: Set the item properties with the boot menu
Initially there may be no items. But when switching to the basic menu
there may well be items already present. Set up their properties so that
they follow the basic-menu theme.

Series-to: concept
Cover-letter:
boot: Separate out the boot-menu style
This series refactors the boot menu so that it is possible to create the
items and then (later) change the style/position to the simple style.
This allows swapping back and forth between a different style and the
simple style.

It also adjusts how the mouse is set up, so that the initial mouse
position is only set up once the video device is known. This allows the
position to be centred on the display, a better starting point than the
top left.
END

Signed-off-by: Simon Glass <sjg@chromium.org>