efi: x86: Use only serial for input
We don't need to enable USB or i8042 for the app. Enable only serial for
now.

Series-to: concept
Cover-letter:
bootctl: Expand bootctl to include a new UI
The current bootctl UI is fairly basic, just supporting a keyboard menu
with text.

Now that expo supports a mouse, add a more interesting UI, with more
graphical elements. Provide a way to switch between this and the simple
UI.

This series also includes some small test improvements, along with a
patch to remove a blob from a bloblist.
END

Signed-off-by: Simon Glass <sjg@chromium.org>