scripts: build-qemu: Allow enabling TKey access
Provide a -T/--tkey option to enable access to a USB-attached Tillitis
TKey.

Series-to: concept
Series-cc: heinrich
Series-links: 1:54
Series-version: 2
Cover-letter:
Bootstage and script enhancements
This series collects together a few improvements to the bootstage
subsystem, principally some new tests.

It also adds support for ACPI FPDT which is a way of passing timing
information to the OS.

A bug fix is provided for mouse clicks with expo.

Finally the series includes some new options for the build-efi script
for networking, spice, etc.
END

Signed-off-by: Simon Glass <sjg@chromium.org>